rename CreateScrum to CRUDScrum
